Update for TV Specials
In this updated version, I've included the term "Special" in the search query to specifically target TV specials. This should help retrieve results that are related to TV specials from nyaa.si. 

Adjustments can be made further based on specific patterns or terms commonly used in the titles of TV specials on nyaa.si.